move routing mode settings to routes/tracks quick menu (fix #12716)
