+++ main/res/layout/map_settings_dialog.xml
@@ -59,84 +59,6 @@
         </com.google.android.material.button.MaterialButtonToggleGroup>
     </RelativeLayout>
 
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp">
-
-        <TextView
-            android:id="@+id/routing_title"
-            style="@style/map_quicksettings_subtitle"
-            android:text="@string/map_routing" />
-
-        <com.google.android.material.button.MaterialButtonToggleGroup
-            android:id="@+id/routing_tooglegroup"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/routing_title"
-            android:layout_marginTop="6dp"
-            app:selectionRequired="true"
-            app:singleSelection="true">
-
-            <Button
-                android:id="@+id/routing_off"
-                style="@style/button_icon_accent"
-                android:layout_margin="0dp"
-                app:icon="@drawable/routing_off" />
-
-            <Button
-                android:id="@+id/routing_straight"
-                style="@style/button_icon_accent"
-                app:icon="@drawable/routing_straight" />
-
-            <Button
-                android:id="@+id/routing_walk"
-                style="@style/button_icon_accent"
-                app:icon="@drawable/routing_walk" />
-
-            <Button
-                android:id="@+id/routing_bike"
-                style="@style/button_icon_accent"
-                app:icon="@drawable/routing_bike" />
-
-            <Button
-                android:id="@+id/routing_car"
-                style="@style/button_icon_accent"
-                app:icon="@drawable/routing_car" />
-        </com.google.android.material.button.MaterialButtonToggleGroup>
-
-        <ImageView
-            android:id="@+id/routing_info"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:src="@drawable/ic_info_blue"
-            android:layout_alignParentRight="true"
-            android:layout_alignBottom="@id/routing_tooglegroup"
-            android:visibility="gone"/>
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:id="@+id/map_settings_autotarget_container"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:visibility="gone">
-
-        <TextView
-            android:id="@+id/map_settings_autotarget_title"
-            style="@style/map_quicksettings_subtitle"
-            android:text="@string/individual_route" />
-
-        <CheckBox
-            android:id="@+id/map_settings_autotarget"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/map_settings_autotarget_title"
-            android:text="@string/map_autotarget_individual_route" />
-    </RelativeLayout>
-
     <View
         android:id="@+id/spacer_bottom"
         android:layout_width="wrap_content"
