+++ main/src/cgeo/geocaching/maps/google/v2/GoogleMapActivity.java
@@ -79,7 +79,7 @@ public class GoogleMapActivity extends AbstractBottomNavigationActivity implemen
     public void onCreate(final Bundle icicle) {
         mapBase.onCreate(icicle);
         routeTrackUtils = new RouteTrackUtils(this, icicle == null ? null : icicle.getBundle(STATE_ROUTETRACKUTILS), mapBase::centerOnPosition,
-            mapBase::clearIndividualRoute, mapBase::reloadIndividualRoute, mapBase::setTrack, mapBase::isTargetSet);
+            mapBase::clearIndividualRoute, mapBase::reloadIndividualRoute, mapBase::setTrack, mapBase::isTargetSet, mapBase::routingModeChanged);
         tracks = new Tracks(routeTrackUtils, mapBase::setTrack);
     }
